Allow to freeze content of a memory view instance
Add a button in the title bar that allows to freeze the current content of the memory view.
If the view is frozen, updates from the debug adapter are ignored and don't overwrite the shown data.
In addition, UI elements that would trigger a data change (core options widget, `load more memory` buttons) are disabled.

Closes #70